fix: 修复加载进度问题

+++ BaseMiniProjectFGUI/BaseProjectMini/Assets/Scripts/Core/Managers/LoadManager.cs
@@ -100,7 +100,7 @@ public class LoadManager : MonoBehaviour
                 else
                 {
                     ++count;
-                    progress?.Invoke(count / maxCount);
+                    progress?.Invoke((float)count / maxCount);
                     if (count >= maxCount)
                     {
                         end?.Invoke(false);
